api-server/requirements.txt
fastapi
uvicorn
python-dotenv
psycopg2-binary
pytest
requests
